segundo cambio

+++ app/Http/Controllers/UserController.php
@@ -6,18 +6,23 @@ use App\Http\Requests\UserCreateRequest;
 use App\Http\Requests\UserEditRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
     public function index()
     {
+        abort_if(Gate::denies('user_index'), 403);
         $users = User::paginate(5);
         return view('users.index', compact('users'));
     }
 
     public function create()
     {
-        return view('users.create');
+        abort_if(Gate::denies('user_create'), 403);
+        $roles = Role::all()->pluck('name', 'id');
+        return view('users.create', compact('roles'));
     }
 
     public function store(UserCreateRequest $request)
@@ -32,18 +37,26 @@ class UserController extends Controller
             + [
             'password' => bcrypt($request->input('password')),
             ]);
+
+        $roles = $request->input('roles', []);
+        $user->syncRoles($roles);
         return redirect()->route('users.show',$user->id)->with('success', 'Usuario creado correctamente');
     }
 
     public function show(User $user)
     {
+        abort_if(Gate::denies('user_show'), 403);
         //$user = User::findOrFail($id);
+        $user->load('roles');
         return view('users.show', compact('user'));
     }
 
     public function edit(User $user)
     {
-        return view('users.edit', compact('user'));
+        abort_if(Gate::denies('user_edit'), 403);
+        $roles = Role::all()->pluck('name', 'id');
+        $user->load('roles');
+        return view('users.edit', compact('user', 'roles'));
     }
 
     public function update(UserEditRequest $request, User $user)
@@ -61,11 +74,18 @@ class UserController extends Controller
         //    $data['password']=bcrypt($request->password);
         //}
         $user->update($data);
+        $roles = $request->input('roles', []);
+        $user->syncRoles($roles);
         return redirect()->route('users.show', $user->id)->with('success', 'Usuario actualizado correctamente');
     }
 
     public function destroy(User $user)
     {
+        abort_if(Gate::denies('user_destroy'), 403);
+        if (auth()->user()->id == $user->id) {
+            return redirect()->route('users.index');
+        }
+
         $user->delete();
         return back()->with('success', 'Usuario eliminado correctamente');
     }
